agents/arbitrage_bot_20260329_002339/requirements.txt
# MCP servers handle all heavy blockchain calls.
# Only lightweight runtime deps needed:
python-dotenv
